+++ TesteSmartHintDomain/Entities/PessoaJuridica.cs
@@ -11,7 +11,7 @@ namespace TesteSmartHint.Domain.Entities
         public PessoaJuridica() { }
 
 
-        public PessoaJuridica(int id, string nome, string email, int telefone, DateTime dtCadastro, bool bloqueado, int? inscricaoEstadual, string CNPJ)
+        public PessoaJuridica(int id, string nome, string email, long telefone, DateTime dtCadastro, bool bloqueado, long? inscricaoEstadual, string CNPJ)
             : base(id, nome, email, telefone, dtCadastro, bloqueado, inscricaoEstadual)
         {
             this.CNPJ = CNPJ;
